Make key.c more like the original archive.
This commit better documents changes by David to switch from stty and
gtty to tcsetattr and tcgetattr.  It's a tiny bit cryptic, but it's
not bad and is less intrusive than the preivous fix.  The problem is
that while there's an sgtty.h that defines the APIs, it doesn't
actually implement them.

Also revert some other whitespace chagne from the original archive.