Bring files into closer alignment with the archive originals.
This commit is half of a change to make it clearer what has been
changed since the files were originally extracted from the archive.
It mostly consists of restoring stupid whitespace changes in the
original source so that diffs more clearly revealed the real changes.

But there are also a number of real changes that have been marked as
either an anachronism or a restoration.  The anchronisms are changes
that were made to allow the original 32-bit compilation and execution,
while the lines marked restoration where the ones that were restored
because the archiving process discarded long lines for some weird
reason.

This commit (in contrast to the next one) contains only the files
where all(?) of the changes were non-controverisial and fairly
straight forward.  It may not actually compile, but the issues should
be minimal if it doesn't.