Bring files into closer alignment with the archive originals.
This commit contains the other half of the changes, a few of which
were more complicated than the previous commit.  In the previous
commit it was 100% whitespace and marking already made changes as
anachronisms or restorations.

In this commit things got a little more complicated because in some
cases, I could get closer to the original archive by declaring
variables in anachronisms.h instead of in the original source files.
In other cases, my restorations were subtly different than the
previous changes because they reduced the number of lines that needed
to be marked.

But the point is that as of this commit diffing the current repo
contents against an original archive (with CRLF line ending
conversions) will result in set of different lines in which *every*
different line is marked with /*anachronism*/ or /*restoration*/.

I've been looking at some of the other early AberMUD distributions
(e.g. abermud89 v1.1 and v1.2 in RCS), and wanted better clarity about
which changes were contemporary versus which were more recent.